# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TideQuest API
  description: API specification for TideQuest - Decentralized identity verification platform
  version: 1.0.0
  contact:
    name: TideQuest Team
    url: https://github.com/TideQuest/ethtokyo2025-sandbox
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/zkp/request:
    post:
      operationId: createZKProofRequest
      summary: Create ZK proof request
      description: Create a zero-knowledge proof request for verification
      tags:
        - ZKP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                providerId:
                  type: string
      responses:
        '200':
          description: Proof request created
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Failed to create proof request
  /api/zkp/verify:
    post:
      operationId: verifyZKProof
      summary: Verify ZK proof
      description: Verify a zero-knowledge proof submission
      tags:
        - ZKP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                provider:
                  type: string
                payload:
                  type: object
      responses:
        '200':
          description: Proof verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
        '400':
          description: Invalid proof format
...
  /api/zkp/steam/proof:
    post:
      operationId: createSteamZKProof
      summary: Create Steam ZK proof
      description: Create a zero-knowledge proof for Steam platform data
      tags:
        - ZKP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                steamId:
                  type: string
                userDataUrl:
                  type: string
                  format: uri
                cookieStr:
                  type: string
                targetAppId:
                  type: string
              required:
                - steamId
                - userDataUrl
                - cookieStr
      responses:
        '200':
          description: Steam proof created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Invalid request parameters
        '500':
          description: Failed to create Steam proof
  /api/zkp/steam/verify:
    post:
      operationId: verifySteamZKProof
      summary: Verify Steam ZK proof
      description: Verify a zero-knowledge proof for Steam platform data
      tags:
        - ZKP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                proof:
                  type: object
                steamId:
                  type: string
                targetAppId:
                  type: string
              required:
                - proof
                - steamId
      responses:
        '200':
          description: Steam proof verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  verified:
                    type: boolean
        '400':
          description: Invalid verification parameters
        '500':
          description: Failed to verify Steam proof
...